splits larger files in raft node
