+++ nodes/raft_state.py
@@ -1,0 +1,8 @@
+from enum import Enum, auto
+
+
+class NodeState(Enum):
+    """Enumeration of Raft node states."""
+    FOLLOWER = auto()
+    CANDIDATE = auto()
+    LEADER = auto()
